Fix regression in SeExpr expression editor
Expressions with $u and $v variables would be considered invalid.

Regression introduced in 59d60f0938f625b02ce544c0949b25a3ac215a4d.
