+++ src/appleseed/foundation/utility/seexpr.h
@@ -32,9 +32,6 @@
 // appleseed.foundation headers.
 #include "foundation/core/concepts/noncopyable.h"
 
-// appleseed.main headers.
-#include "main/dllsymbol.h"
-
 // Boost headers.
 #include "boost/regex.hpp"
 
@@ -45,28 +42,6 @@
 
 namespace foundation
 {
-
-//
-// Checks the validity of a SeExpr expression.
-//
-
-class APPLESEED_DLLSYMBOL SeExprValidator
-  : public NonCopyable
-{
-  public:
-    explicit SeExprValidator(const char* expression);
-
-    ~SeExprValidator();
-
-    bool is_valid() const;
-
-    const char* get_parse_error() const;
-
-  private:
-    struct Impl;
-    Impl* impl;
-};
-
 
 //
 // Extracts all file paths from a valid SeExpr expression.
